move the context handler from run.py to runtimes
